devtools: update filter criteria for backports
It was important in the past to select anything with "fix" in the
headline, but recently more often created false positives and work
to sort things out than identifying many helpful patches.

The community and processes around DPDK matured enough that developers
(rightfully) expect to rely on "Fixes:" or "stable@" marking for
backports.

Therefore do no more include patches that just include the word fix
in the backport candidate list.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Acked-by: Luca Boccassi <bluca@debian.org>
Acked-by: Kevin Traynor <ktraynor@redhat.com>
Acked-by: Xueming Li <xuemingl@nvidia.com>