devtools: check commit tag sequence
This change to log checking procedure ensures that certain
tags are in proper order.
The order of tags is as follows:
 * Coverity issue
 * Bugzilla ID
 * Fixes
 * Cc
 * <BLANK LINE>
 * Reported-by
 * Suggested-by
 + Signed-off-by
 * Acked-by
 * Reviewed-by
 * Tested-by
where:
 * => 0 or more than one instance possible
 + => more than once instance possible
In order to satisfy the above requirements an extra check
is performed for obligatory tags.

Note: The last block of tags ending with "-by"
should be ordered in chronological order.

Signed-off-by: Jakub Palider <jpalider@marvell.com>
Signed-off-by: Thomas Monjalon <thomas@monjalon.net>