eal: use dedicated PRNG for non-lcore threads
Prior to this change, unregistered non-EAL threads shared a PRNG
instance with the main lcore. The main lcore may well be used for fast
path processing, potentially making rte_rand() calls in the
process. It should not need to synchronize with control threads.

With this change, all unregistered non-EAL threads share one dedicated
PRNG instance.

The API documentation is updated to use the proper terminology when
referring to threads equipped with an lcore id.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>
Acked-by: Morten Brørup <mb@smartsharesystems.com>