net/mlx5/hws: support ibv context shared with local one
The idea is to have a shared ibv_context that all the resources
allocated on it (FT + TIR are exceptions)
Whenever a resource is created locally an alias object to that resource
allocated and used in the other context.
The connections between the resources are done according to each type of
the resource, to the original resource, or its alias resource.

Signed-off-by: Erez Shitrit <erezsh@nvidia.com>
Reviewed-by: Alex Vesker <valex@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>