usertools/telemetry: add file prefix argument
Currently, the file prefix for DPDK runtime directory is 'rte' which was
fixed in the telemetry client script. The user had to modify the prefix
each time to run this script if the file prefix of application isn't 'rte'.
So this patch adds an optional argument for the file prefix, like,
"./dpdk-telemetry-client.py -f rte_xxx sock_path"

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>