net/nfp: update macro definitions in Tx flag
In firmware, bit 0 and bit 1 in Tx descriptor flag are defined
as "PCIE_DESC_TX_ENCAP" and "PCIE_DESC_TX_O_IP4_CSUM". Meanwhile,
bit 0 is used to mark tunnel packets and bit 1 is used to set
outer ip cusm offload. Thus, update the macro definitions as
same as the ones in firmware.

Signed-off-by: Fei Qin <fei.qin@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>