net/bnxt: remove compile-time option for IEEE 1588
The drivers are supposed to enable/disable features dynamically
instead of compile time options.
Remove the compile time option RTE_LIBRTE_IEEE1588.

This patch is an improved and bnxt specific version of the previous
submission which tried to remove the option tree wide, but was
completely disabling for bnxt PMD.

http://patchwork.dpdk.org/project/dpdk/patch/20230203132810.14187-1-thomas@monjalon.net/

Also add a devarg to enable/disable the setting,
invoked as for ex: "-a 000:0b:0d.0,ieee-1588=1".

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>