net/sfc: replace PMD log type
The static LOGTYPE PMD will go away in the future.
Replace its use in the sfc driver with the same static
type as is already used in RTE_LOG_REGISTER macro.

These drivers are trying to do things that other drivers are not
in creating log types. IMHO it should follow the other driver
usage pattern. But this patch leaves the special code.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>