common/mlx5: add CQE validity iteration count
The validity iteration count replaces the functionality of the owner bit
in terms of indicating that a new CQE was written to buffer.
On iteration=k on the CQ buffer, only entries with the iteration_count=k
should be treated as new CQEs or mini CQE arrays.
The validity iteration count is used when the Enhanced CQE compression
is selected. Add this CQE field and the method to check it.

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>