test/crypto: improve readability of ZUC-256 cases
There were some ZUC-256 tests for hash generate and hash verify
mixed in with the ZUC-128 tests,
with no indication in the test name that they were for ZUC256.

These are now grouped with other ZUC-256 tests, and
all ZUC-256 hash test names are modified to be more uniform, and
indicate the digest length as this can vary with ZUC-256.

Signed-off-by: Ciara Power <ciara.power@intel.com>
Acked-by: Brian Dooley <brian.dooley@intel.com>
Acked-by: Tejasree Kondoj <ktejasree@marvell.com>
Acked-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>