dts: add test results module
The module stores the results and errors from all executions, build
targets, test suites and test cases.

The result consist of the result of the setup and the teardown of each
testing stage (listed above) and the results of the inner stages. The
innermost stage is the case, which also contains the result of test case
itself.

The modules also produces a brief overview of the results and the
number of executed tests.

It also finds the proper return code to exit with from among the stored
errors.

Signed-off-by: Juraj Linkeš <juraj.linkes@pantheon.tech>
Tested-by: Bruce Richardson <bruce.richardson@intel.com>
Tested-by: Chenyu Huang <chenyux.huang@intel.com>
Tested-by: Patrick Robb <probb@iol.unh.edu>