common/sfc_efx/base: add MAE mark reset action
Previously, DPDK sfc driver received support for tunnel
offload. In it, MAE needs to set intermediate mark from
an outer rule (OR) recirculation ID in order to help
the driver identify packets that hit the OR but miss
on action rule (AR) lookup. But, for packets that do
hit an AR, the driver wants to reset this mark so
that the end receiver of traffic does not see it.

The driver has a call to request such mark reset,
but it does not work as it comes via the regular
mark populate API, which must not be invoked
after final delivery action has been added.

Provide a suitable dedicated API for that.

Fixes: 3a73dcfdb255 ("common/sfc_efx/base: match on recirc ID in action rules")
Cc: stable@dpdk.org

Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Reviewed-by: Andy Moreton <amoreton@xilinx.com>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>