vhost: fix madvise IOTLB entries pages overlap check
At removal time, when testing whether the IOTLB entry has
shared pages with the previous and next entries in the
cache, it checks whether the start address of the entry to
be removed is on the same page as the start address of the
next entry in the cache.

This is not correct, as an entry could cover several page
so the end address of the entry to be remove should be
used. This patch address this issue.

Note there is another issue not fixed by this patch, but
delayed to next release given its minor impact and the
complexity of the fix it requires. If a removed IOTLB entry
is spanned on several pages and one of the pages is shared
with another entry, all the pages will remain as DODUMP
while only the shared page should be. It would result in
non-shared pages to be part of the coredump while it would
not be needed.

Fixes: dea092d0addb ("vhost: fix madvise arguments alignment")

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Reviewed-by: Chenbo Xia <chenbo.xia@intel.com>